sinan-core/test_mlx.py
#!/usr/bin/env python3
"""测试 MLX 后端是否能正常加载和运行"""
import os
import sys
import platform
from pathlib import Path

# 设置 Hugging Face Token 加速下载（从环境变量读取，避免硬编码）
os.environ["HF_TOKEN"] = os.environ.get("HF_TOKEN", "")

# 添加 src 到路径
sys.path.insert(0, str(Path(__file__).parent / "src"))


def check_platform():
    """检查是否为 Apple Silicon Mac"""
    print(f"平台信息:")
    print(f"  系统: {platform.system()}")
    print(f"  架构: {platform.machine()}")
    print(f"  处理器: {platform.processor()}")

    is_apple_silicon = (
        platform.system() == "Darwin" and
        platform.machine() in ["arm64", "aarch64"]
    )

    if not is_apple_silicon:
        print("\n⚠️  警告: 当前不是 Apple Silicon Mac，MLX 后端不可用")
        return False

    print("\n✅ 检测到 Apple Silicon Mac")
    return True


def check_mlx_installed():
    """检查 mlx-lm 是否已安装"""
    try:
        import mlx_lm
        print(f"✅ mlx-lm 已安装")
        return True
    except ImportError:
        print("❌ mlx-lm 未安装")
        print("   请运行: uv add mlx-lm")
        return False


def test_mlx_backend():
    """测试 MLX 后端"""
    print("\n" + "="*50)
    print("测试 MLX 后端")
    print("="*50)

    # 1. 检查平台
    if not check_platform():
        return False

    # 2. 检查依赖
    if not check_mlx_installed():
        return False

    # 3. 尝试初始化 MLX 后端
    print("\n正在初始化 MLX 后端...")
    model_path = "/Users/void/.cache/sinan/mlx-models/MAI-UI-8B-bf16"
    print(f"模型: {model_path}")
    print("-"*50)

    try:
        from sinan_core.vision.mlx_backend import MLXBackend

        backend = MLXBackend(model_name=model_path)
        success = backend.initialize()

        if success:
            print("\n✅ MLX 后端初始化成功!")
            print(f"   处理器已加载: {backend.processor is not None}")
            print(f"   模型已加载: {backend.model is not None}")
            return True
        else:
            print("\n❌ MLX 后端初始化失败")
            return False

    except Exception as e:
        print(f"\n❌ 初始化过程中出错: {type(e).__name__}: {e}")
        import traceback
        traceback.print_exc()
        return False


def test_vision_agent():
    """测试 VisionAgent 自动选择 MLX 后端"""
    print("\n" + "="*50)
    print("测试 VisionAgent 自动选择后端")
    print("="*50)

    try:
        from sinan_core.vision import VisionAgent

        model_path = "/Users/void/.cache/sinan/mlx-models/MAI-UI-8B-bf16"
        agent = VisionAgent(mlx_model=model_path)
        success = agent.initialize()

        if success:
            print(f"\n✅ VisionAgent 初始化成功!")
            print(f"   使用的后端: {agent.backend_name}")

            if agent.backend_name == "mlx":
                print("   🎉 成功使用 MLX 后端!")
            else:
                print(f"   注意: 当前使用的是 {agent.backend_name} 后端，不是 MLX")
            return True
        else:
            print("\n❌ VisionAgent 初始化失败")
            return False

    except Exception as e:
        print(f"\n❌ 测试过程中出错: {e}")
        import traceback
        traceback.print_exc()
        return False


def test_inference():
    """测试实际的图像推理"""
    print("\n" + "="*50)
    print("测试图像推理")
    print("="*50)

    try:
        from sinan_core.vision import VisionAgent
        from PIL import Image

        model_path = "/Users/void/.cache/sinan/mlx-models/MAI-UI-8B-bf16"
        agent = VisionAgent(mlx_model=model_path)
        if not agent.initialize():
            print("❌ VisionAgent 初始化失败")
            return False

        # 创建一个简单的测试图像（红色背景上的蓝色方块）
        print("创建测试图像...")
        img = Image.new('RGB', (800, 600), color='red')
        from PIL import ImageDraw
        draw = ImageDraw.Draw(img)
        draw.rectangle([300, 200, 500, 400], fill='blue')

        print("测试指令: '蓝色方块'")
        result = agent.detect_element(img, "蓝色方块")

        # 获取原始响应用于调试
        from sinan_core.vision.mlx_backend import MLXBackend
        backend = MLXBackend(model_name=model_path)
        backend.initialize()
        raw_response = backend.get_raw_response(img, "蓝色方块")
        print(f"   原始响应: {raw_response}")

        if result:
            print(f"✅ 检测到元素!")
            print(f"   BBox: {result['bbox']}")
            print(f"   中心点: {result['center']}")
            return True
        else:
            print("⚠️  未检测到元素（这在测试图像上是正常的）")
            return True  # 推理流程正常，只是没检测到

    except Exception as e:
        print(f"❌ 推理测试失败: {e}")
        import traceback
        traceback.print_exc()
        return False


def main():
    print("MLX 后端测试脚本")
    print("="*50)

    # 检查 HF_TOKEN
    if not os.environ.get("HF_TOKEN"):
        print("\n⚠️  警告: 未设置 HF_TOKEN 环境变量")
        print("   如需下载模型，请先设置:")
        print("   export HF_TOKEN=your_token_here")
        print()

    # 测试 MLX 后端
    mlx_ok = test_mlx_backend()

    # 测试 VisionAgent
    agent_ok = test_vision_agent()

    # 测试实际推理
    inference_ok = test_inference() if (mlx_ok and agent_ok) else False

    # 总结
    print("\n" + "="*50)
    print("测试结果总结")
    print("="*50)
    print(f"MLX 后端初始化: {'✅ 通过' if mlx_ok else '❌ 失败'}")
    print(f"VisionAgent 初始化: {'✅ 通过' if agent_ok else '❌ 失败'}")
    print(f"图像推理测试: {'✅ 通过' if inference_ok else '❌ 失败'}")

    if mlx_ok and agent_ok:
        print("\n🎉 MLX 后端可以正常使用!")
        return 0
    else:
        print("\n⚠️  部分测试未通过，请检查错误信息")
        return 1


if __name__ == "__main__":
    sys.exit(main())
