sinan-core/pyproject.toml
[project]
name = "sinan-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adb-shell>=0.4.4",
    "fastapi>=0.128.0",
    "paddleocr>=3.3.3",
    "pillow>=12.1.0",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "openai>=1.0.0",
    "transformers>=4.57.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "mlx-vlm>=0.2.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "modelscope>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sinan_core"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
